reorg files and clean up
