+++ client/src/components/app.jsx
@@ -8,7 +8,7 @@ import {
 } from 'react-router-dom';
 import './common/fontAwesomeIcons';
 
-import Banner from './common/banner.jsx';
+import NavBar from './common/navBar.jsx';
 import Wrapper from './Wrapper.jsx';
 
 class App extends React.Component {
@@ -20,7 +20,7 @@ class App extends React.Component {
   render() {
     return (
       <Router>
-        <Banner />
+        <NavBar />
         <Switch>
           <Route exact={false} path='/product/:productId' component={Wrapper} {...this.props}>
             {/* <OverviewContainer />
