+++ client/src/components/common/commonUI.module.css
@@ -8,7 +8,7 @@
   margin-left: 30px;
 }
 
-.banner {
+.nav-bar {
   background-color: #92a4b3;
   width: 100%;
   height: 100px;
