+++ client/src/components/ratingsAndReviews/ratings/productBreakdown.jsx
@@ -1,11 +1,11 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import { connect } from 'react-redux';
-import CSS from './productReviews.module.css';
+import CSS from '../ratingsAndReviews.module.css';
 
-const ProductBreakdown = (props) => (
+const ProductBreakdown = ({ reviewMeta }) => (
     <div className={CSS['product-breakdown']}>
-      {props.reviewMeta.characteristics && Object.keys(props.reviewMeta.characteristics)
+      {reviewMeta.characteristics && Object.keys(reviewMeta.characteristics)
         .map((key) => <div key={key}>{key}: ----*----</div>)}
     </div>
 );
