+++ client/src/components/ratingsAndReviews/ratings/ratings.jsx
@@ -1,31 +1,30 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import { connect } from 'react-redux';
-import StarRating from '../common/starRating.jsx';
+import StarRating from '../../common/starRating.jsx';
 import StarBreakdown from './starBreakdown.jsx';
 import ProductBreakdown from './productBreakdown.jsx';
-import CSS from './productReviews.module.css';
+import CSS from '../ratingsAndReviews.module.css';
 
-const Ratings = (props) => (
-  <div>
-    {props.reviews.length === 0
+const Ratings = ({ reviews, reviewMeta }) => (
+  <div className={CSS['ratings-container']}>
+    {reviews.length === 0
       ? <>No Reviews Available</>
       : <>
-          {props.reviewMeta.ratingScore && (
+          {reviewMeta.ratingScore && (
             <div className={CSS['rating-summary']}>
-              <div className={CSS['rating-score']}>{props.reviewMeta.ratingScore}</div>
+              <div className={CSS['rating-score']}>{reviewMeta.ratingScore}</div>
               <StarRating />
             </div>
           )}
-          {props.reviewMeta.recommendationRate && (
-            <div className={CSS['rating-recommendation']}>{props.reviewMeta.recommendationRate !== 'NaN' ? props.reviewMeta.recommendationRate : 0}%
+          {reviewMeta.recommendationRate && (
+            <div className={CSS['rating-recommendation']}>{reviewMeta.recommendationRate !== 'NaN' ? reviewMeta.recommendationRate : 0}%
             of reviews recommend this product</div>
           )}
           <StarBreakdown />
           <ProductBreakdown />
         </>
     }
-
   </div>
 );
