+++ client/src/components/ratingsAndReviews/test/ratings.test.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
 import testStore from '../../../fixtures/testStore';
-import Ratings from '../ratings.jsx';
+import Ratings from '../ratings/ratings.jsx';
 import '../../common/fontAwesomeIcons';
 
 describe('Ratings', () => {
