+++ client/src/components/ratingsAndReviews/test/starBreakdown.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import StarBreakdown from '../starBreakdown.jsx';
+import StarBreakdown from '../ratings/starBreakdown.jsx';
 import testStore from '../../../fixtures/testStore';
 import '../../common/fontAwesomeIcons';
