+++ client/src/reducers/rootReducer.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
-import reviews from './productReviews/reviewsReducer';
-import filters from './productReviews/filtersReducer';
-import loadedReviews from './productReviews/loadedReviewsReducer';
-import remainingReviews from './productReviews/remainingReviewsReducer';
-import reviewMeta from './productReviews/reviewMetaReducer';
+import reviews from './ratingsAndReviews/reviewsReducer';
+import filters from './ratingsAndReviews/filtersReducer';
+import loadedReviews from './ratingsAndReviews/loadedReviewsReducer';
+import remainingReviews from './ratingsAndReviews/remainingReviewsReducer';
+import reviewMeta from './ratingsAndReviews/reviewMetaReducer';
 import currentProduct from './productOverview/currentProduct';
 import styleList from './productOverview/styleList';
 import currentStyle from './productOverview/currentStyle';
